temp commit

+++ lib/Diff_drive_unicycle/Diff_drive_unicycle.cpp
@@ -3,84 +3,84 @@
 Diff_drive_unicycle::Diff_drive_unicycle(){
 }
 
-void Diff_drive_unicycle::set_param(float r_, float L_){
+void Diff_drive_unicycle::set_param(double r_, double L_){
     r = r_;
     L = L_;
 }
 
-void Diff_drive_unicycle::set_v_max(float V_max_){
+void Diff_drive_unicycle::set_v_max(double V_max_){
     V_max = V_max_;
     w_lr_max = V_max/r;
     W_max = 2.0*V_max/L;
 }
 
-void Diff_drive_unicycle::set_w_max(float W_max_){
+void Diff_drive_unicycle::set_w_max(double W_max_){
     W_max = W_max_;
     V_max = W_max*L/2.0;
     w_lr_max = V_max/r;
 }
 
-float Diff_drive_unicycle::get_v_max(){
+double Diff_drive_unicycle::get_v_max(){
     return V_max;
 }
 
-float Diff_drive_unicycle::get_w_max(){
+double Diff_drive_unicycle::get_w_max(){
     return W_max;
 }
 
-float Diff_drive_unicycle::get_wlr_max(){
+double Diff_drive_unicycle::get_wlr_max(){
     return w_lr_max;
 }
 
-void Diff_drive_unicycle::set_r(float r_){
+void Diff_drive_unicycle::set_r(double r_){
     r = r_;
 }
 
-void Diff_drive_unicycle::set_L(float L_){
+void Diff_drive_unicycle::set_L(double L_){
     L = L_;
 }
 
-float Diff_drive_unicycle::get_r(){
+double Diff_drive_unicycle::get_r(){
     return r;
 }
 
-float Diff_drive_unicycle::get_L(){
+double Diff_drive_unicycle::get_L(){
     return L;
 }
 
-void Diff_drive_unicycle::uni2ddr(float Vc_, float Wc_, float* wr, float* wl){
-    float Vc = Vc_;
-    float Wc = Wc_;
+void Diff_drive_unicycle::uni2ddr(double Vc_, double Wc_, double* wr, double* wl){
+    double Vc = Vc_;
+    double Wc = Wc_;
     update_domain_vw(Vc_, Wc_, &Vc, &Wc);
     
     *wr = (Vc + Wc*L*0.5) / r;
     *wl = (Vc - Wc*L*0.5) / r;
 }
 
-float Diff_drive_unicycle::get_wr(float Vc, float Wc){
+double Diff_drive_unicycle::get_wr(double Vc, double Wc){
     return (Vc + Wc*L*0.5) / r;
 }
 
-float Diff_drive_unicycle::get_wl(float Vc, float Wc){
+double Diff_drive_unicycle::get_wl(double Vc, double Wc){
     return (Vc - Wc*L*0.5) / r;
 }
 
-void Diff_drive_unicycle::ddr2uni(float wr, float wl, float* Vc, float* Wc){
+void Diff_drive_unicycle::ddr2uni(double wr, double wl, double* Vc, double* Wc){
     *Vc = (wr + wl)*r*0.5;
     *Wc = (wr - wl)*r/L;
 }
 
-float Diff_drive_unicycle::get_Vc(float wr, float wl){
+double Diff_drive_unicycle::get_Vc(double wr, double wl){
     return (wr + wl)*r*0.5;
 }
 
-float Diff_drive_unicycle::get_Wc(float wr, float wl){
+double Diff_drive_unicycle::get_Wc(double wr, double wl){
     return (wr - wl)*r/L;
 }
 
-void Diff_drive_unicycle::update_domain_vw(float V_in, float W_in, float* V_out, float* W_out){
-    float V_n = V_in;
-    float W_n = W_in;
+void Diff_drive_unicycle::update_domain_vw(double V_in, double W_in, double* V_out, double* W_out){
+    double V_n = V_in;
+    double W_n = W_in;
 
     int region = 0;
